feat(fixtures): Kotlin-based multi-repo sandbox + CLI
