+++ maple/src/main/kotlin/com/tazzledazzle/maple/fixtures/FixtureDriver.kt
@@ -1,0 +1,6 @@
+package com.tazzledazzle.maple.fixtures
+
+interface FixtureDriver {
+    fun create(spec: FixturesSpec): List<BomEntry>
+    fun destroy(entries: List<BomEntry>, spec: FixturesSpec)
+}
